+++ src/infrastructure/applications/rest/RestApplication.ts
@@ -1,8 +1,8 @@
 import {NestFactory, Reflector} from '@nestjs/core';
-import {json, urlencoded} from 'body-parser';
+import {NestExpressApplication} from '@nestjs/platform-express';
 import * as cookieParser from 'cookie-parser';
 import {DocumentBuilder, SwaggerModule} from '@nestjs/swagger';
-import {INestApplication, VersioningType} from '@nestjs/common';
+import {VersioningType} from '@nestjs/common';
 import * as Sentry from '@sentry/nestjs';
 import {SentryExceptionFilter} from './SentryExceptionFilter';
 import {SchemaSerializer} from './SchemaSerializer';
@@ -22,7 +22,7 @@ export class RestApplication extends BaseApplication {
      * An instance of an application built with NestJS.
      * @protected
      */
-    protected _app: INestApplication;
+    protected _app: NestExpressApplication;
 
     /**
      * The class of the application module (default is `AppModule`).
@@ -176,9 +176,13 @@ export class RestApplication extends BaseApplication {
      * @protected
      */
     protected initSettings() {
-        this._app.use(json({limit: this._config.requestSizeLimit}));
-        this._app.use(urlencoded({extended: true,
-            limit: this._config.requestSizeLimit}));
+        this._app.useBodyParser('json', {
+            limit: this._config.requestSizeLimit,
+        });
+        this._app.useBodyParser('urlencoded', {
+            extended: true,
+            limit: this._config.requestSizeLimit,
+        });
     }
 
     /**
@@ -186,7 +190,7 @@ export class RestApplication extends BaseApplication {
      * @protected
      */
     protected async createApp() {
-        this._app = await NestFactory.create(this._moduleClass, {
+        this._app = await NestFactory.create<NestExpressApplication>(this._moduleClass, {
             logger: this._config.loggerLevels,
         });
     }